AHS-Freshman-Debates/requirements.txt
Django>=1.5
python-social-auth
django-import-export
